Created Our Individual Scenes
These can be found under Assets -> Scenes -> Our Individual Scenes. These scenes are for if we need to go in and add, experiment, or change things that have to do with our own assignments, so we aren't stepping on each other's toes. The Main Game Scenes Folder is for scenes that are made with the finished features, and make up the actual Game.